Funcionando y listas las funciones de la parte 1

+++ Libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: alejjime <alejjime@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/03 15:40:33 by alejjime          #+#    #+#             */
-/*   Updated: 2024/12/11 21:08:23 by alejjime         ###   ########.fr       */
+/*   Updated: 2024/12/27 18:11:41 by alejjime         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ it until find the '\0' value.
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
